requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
sqlalchemy>=2.0.40,<3
aiosqlite>=0.21,<1
pydantic>=2.12,<3
python-jose[cryptography]>=3.4,<4
passlib[bcrypt]>=1.7.4,<2
bcrypt>=4.0.1,<5
python-dotenv>=1.1,<2
pytest>=8.3,<9
pytest-asyncio>=0.26,<1
ruff>=0.11,<1
